refactor(indexer/handlers): extract fpmm/state-sync.ts (UpdateReserves + Rebalanced) (PR-S3)
Third slice of the handlers/fpmm.ts split (after PR-S1 #297 factory.ts
and PR-S2 #298 liquidity.ts). Move FPMM.UpdateReserves and FPMM.Rebalanced
into the new indexer-envio/src/handlers/fpmm/state-sync.ts (~360 lines).

Critical invariants preserved verbatim:
- Double-write health pattern: upsertPool → recordHealthSample → Pool.set
- `-2` sentinel skip on `rebalanceReward` (fetchRebalanceIncentiveAtBlock
  returning -2n means the getter doesn't exist on this contract; do not
  clobber the existing pool field)
- Promise.all / Promise.allSettled concurrency patterns (load-bearing
  for handler latency)

EventHandlers.ts is updated to import the new module so Envio sees the
side-effect handler registrations at module load.

handlers/fpmm.ts: 678 → 331 lines.

Pure extraction. 0 tests pass (pre-existing baseline on this branch).
